shinano: overlay: WiFi chipset has background scan support
Boolean indicating whether the wifi chipset supports
background scanning mechanism.
This mechanism allows the host to remain in suspend
state and the dongle to actively scan and wake the
host when a configured SSID is detected by the dongle.
This chipset capability can provide power savings
when wifi needs to be always kept on

Signed-off-by: Humberto Borba <humberos@gmail.com>