Add a simple PowerHAL for native double-tap-to-wake
We dropped the QCOM PowerHAL because it relies on too many
proprietary blobs. Instead we have a kernel solution, but
several features still depend on the presence of a PowerHAL,
so Android is actually not happy about it.

Marshmallow introduced native double-tap-to-wake support,
which is built into the PowerHAL. This commit enables DTTW
and adds a very simple PowerHAL to support this feature.

Hopefully we can later extend this when our kernel driver
can accept changes from userspace.

Signed-off-by: Adam Farden <adam@farden.cz>