power: add some PowerHAL enhancements
When the device is asleep we don't need to scale up the cpu
frequencies so we can tune the governor to make it twice as hard
to scale up and twice as easy to scale down.

We do this by adding profiles. the 'power saving' profile can
be used when the device is asleep, or when Android enters a low
power mode. This profile also lowers the 'hispeed' from 1190MHz
to 960MHz.

The PowerHAL can also briefly boost frequencies to the 'hispeed'
frequency when the user interacts with the device. The boost is
only for 80ms but this dramatically improves responsiveness,
including 20% quicker wake from sleep.

Also fix some warnings, and the make sure we don't leave 'power
saving' mode when waking if 'low power' mode is active.

Signed-off-by: Adam Farden <adam@farden.cz>