sony: shinano: bluetooth: Update the BT FW init parameters
Update the BT parameters of LPM_IDLE_THRESHOLD and LPM_HC_IDLE_THRESHOLD to 24.
Each count of threshold is 12.5ms equivalent with.
The new value (24) is equivalent with 300ms.

LPM_IDLE_THRESHOLD
---------------------
In sleep mode 1, this is the number of firmware loops executed with no
activity before the Host wake line is deasserted. Activity includes HCI
traffic excluding certain sleep mode commands and the presence of SCO
connections if the "Allow Host Sleep During SCO" flag is not set to 1.
Each count of this parameter is roughly equivalent to 300ms or 25ms.

LPM_HC_IDLE_THRESHOLD
---------------------
This is the number of firmware loops executed with no activity
before the HC is considered idle. Depending on the mode, HC may then attempt
to sleep.
Activity includes HCI traffic excluding certain sleep mode commands and the
presence of ACL/SCO connections.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: I789d52822a4ed11a56a43a668e0aa9493bbded11